+++ rfx/runners/uniform.py
@@ -534,6 +534,7 @@ def run_uniform(
         waveguide_ports=waveguide_ports_result,
         waveguide_sparams=waveguide_sparams_result,
         snapshots=sim_result.snapshots,
+        grid=grid,
         dt=grid.dt,
         freq_range=(sim._freq_max / 10, sim._freq_max, sim._boundary),
     )